Implement basic single-client functionality
So...this commit is huge because I kind of went on a coding tear and
implemented a whole bunch of stuff all at once. Not the best strategy,
but since this is an experimental project I don't consider it too huge
of a sin. Here's a list of the basic stuff I implemented:

* Got rid of the old VB project, replaced it with WPF/C#
* Created two models: Conversation and Message (not POD enough)
* Created a Filesystem utility for interacting with the OS
* Implemented a backend for conversation data via a HTML file
* Added dependencies on HTML Agility Project and Extended WPF Toolkit

On the plus side, single-client communication works! It's file-based,
for now, and my next step is to get it listening on a port for incoming
messages from both the client and, well, another client. It's also super
decoupled and the code is really nice! (except for some messiness in the
models, but I'll fix those later)

From now on I'll be committing more frequently to avoid massive commits.